+++ src/libsquey/tests/mapping_host_default.cpp
@@ -65,11 +65,11 @@ int main()
 
 		// Check IP are in 0 -> 2*31
 		uint32_t map_ip = mapped.get_column(PVCol(0)).to_core_array<uint32_t>()[i];
-		PV_ASSERT_VALID(map_ip < (1UL << 31));
+		PV_ASSERT_VALID(map_ip < ((uint32_t)1 << 31));
 
 		// Check str are in 2*31 -> 2**32
 		uint32_t map_str = mapped.get_column(PVCol(1)).to_core_array<uint32_t>()[i];
-		PV_ASSERT_VALID(map_str >= (1UL << 31));
+		PV_ASSERT_VALID(map_str >= (((uint32_t)1 << 31)));
 	}
 #else
 	(void)mapped;
